src/requirements.txt
faker.py==1.0
